Extract verifier alliance setup into a helper mod
